Formatting landing page using css

+++ SVG_Restaurants/Controllers/CustomersController.cs
@@ -91,7 +91,7 @@ namespace SVG_Restaurants.Controllers
             {
                 _context.Add(customer);
                 await _context.SaveChangesAsync();
-                return RedirectToAction("Login", "Customers");
+                return RedirectToAction(nameof(Index));
             }
             return View(customer);
         }
@@ -168,25 +168,20 @@ namespace SVG_Restaurants.Controllers
         // POST: Customers/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int CustomerID)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            try
+            if (_context.Customers == null)
             {
-                // Attempt to find the customer with the specified CustomerId
-                var customer = _context.Customers.Single(c => c.CustomerId == CustomerID);
-
-                // Remove the customer entity
+                return Problem("Entity set 'SGVContext.Customers'  is null.");
+            }
+            var customer = await _context.Customers.FindAsync(id);
+            if (customer != null)
+            {
                 _context.Customers.Remove(customer);
-
-                // Save the changes to the database
-                await _context.SaveChangesAsync();
-
-                return RedirectToAction("Index", "Home");
             }
-            catch (InvalidOperationException)
-            {
-                return NotFound(); // Handle the case where the customer is not found.
-            }
+            
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
         }
 
         private bool CustomerExists(int id)
